Show every mail in the box, not just the first one
NoteDto.ReadCount was declared as uint but both clients read two bytes there.
Every element of the array came out two bytes long and the drift piled up, so
the first mail was parsed fine and the rest arrived with their strings empty.
With a single mail in the box it never showed.

The drift is visible in the read request: the client asked for 1376512, which
is 0x00150100 - the 00 01 are the DaysLeft and Unk8 of the previous element
and only then comes the id 21. Both clients walk the array with a stride of
0x52, confirmed on the S8 and the S10 binary.

Clan mails also need a different message type per client: S10 reads type 1
fine, while on S8 that type makes the client pack the id together with the
type and the popup comes back empty, so it goes out as type 2 there.

+++ src/GameServer/Network/GameServer.cs
@@ -417,7 +417,7 @@ namespace Santana.Network
                 .Function(dest => dest.Unk1, src => (ulong)(src.Id & 0xFFFFFFFF))
                 .Function(dest => dest.Receiver, src => (ulong)(src.Id >> 32))
                 .Member(dest => dest.Title, src => src.Title)
-                .Function(dest => dest.ReadCount, src => src.IsNew ? 0 : 1)
+                .Function(dest => dest.ReadCount, src => src.IsNew ? (ushort)0 : (ushort)1)
                 .Function(dest => dest.DaysLeft,
                     src => DateTimeOffset.Now < src.Expires
                         ? (byte)System.Math.Max(0, System.Math.Min(255,
